third_party/blink/web_tests/platform/generic/fast/js/regexp-unicode-overflow-expected.txt

This test checks for a regression against REGRESSION: Apparent WebKit JavaScript memory smasher when submitting comment to iWeb site (crashes in kjs_pcre_compile2). If it fails, it may crash.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS new RegExp("[k ]", "i").toString() is /[k ]/i.toString()
PASS new RegExp("[k’]", "i").toString() is /[k’]/i.toString()
PASS new RegExp("[k“]", "i").toString() is /[k“]/i.toString()
PASS new RegExp("[k”]", "i").toString() is /[k”]/i.toString()
PASS successfullyParsed is true

TEST COMPLETE
